refactor(types): group office domain under common/types/office
