+++ packages/desktop/src/common/types/office/preview.ts
