+++ packages/desktop/src/renderer/hooks/file/usePreviewLauncher.ts
@@ -6,7 +6,7 @@
 
 import { ipcBridge } from '@/common';
 import { joinPath } from '@/common/chat/chatLib';
-import type { PreviewContentType } from '@/common/types/preview';
+import type { PreviewContentType } from '@/common/types/office/preview';
 import { useConversationContextSafe } from '@/renderer/hooks/context/ConversationContext';
 import { usePreviewContext } from '@/renderer/pages/conversation/Preview';
 import {
