+++ packages/desktop/src/renderer/pages/conversation/Preview/components/PreviewPanel/PreviewHistoryDropdown.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { PreviewHistoryTarget, PreviewSnapshotInfo } from '@/common/types/preview';
+import type { PreviewHistoryTarget, PreviewSnapshotInfo } from '@/common/types/office/preview';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
