+++ packages/desktop/src/renderer/utils/emitter.ts
@@ -8,7 +8,7 @@ import EventEmitter from 'eventemitter3';
 import type { DependencyList } from 'react';
 import { useEffect } from 'react';
 import type { FileOrFolderItem } from '@/renderer/utils/file/fileTypes';
-import type { PreviewContentType } from '@/common/types/preview';
+import type { PreviewContentType } from '@/common/types/office/preview';
 
 export type ReplyQuote = {
   messageId: string;
